requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-core
faiss-cpu
sentence-transformers
numpy
# psycopg2

# -e .
